(core) Renamed some modules interacting with the "runtime system"
I've used "runtime system" to denote the part of the "Ciao system" that
implements parts of the execution model and its behaviour. That is a
currently accepted notion of "language runtime".

Note that this is not exactly the same as the Ciao engine, since the
runtime may include some basic libraries and exclude the bytecode
emulator (i.e., compiling to C still requires a "runtime system").

Src-commit: e912e70898994a2a4e3b6b482387f73bf7ccc13e