(core) Removed open_null_stream/1, added system:dev_null/1
The open_null_stream/1 predicate was not used in any Ciao code.

It opened a write stream to the /dev/null device, which only works in
POSIX. I've replaced it with a dev_null/1 predicate which maps to the
file name '/dev/null' (POSIX) or 'nul' (Windows), so that open/3 can be
used to open a stream in read/write mode. This is the solution adopted
by other languages.

The process.pl library now uses this predicate in stdin/stdout
redirections to null.

Src-commit: dea1189725fbd4d955995348611a2e7abba000fa