Fixing doc(nodoc,assertions) in documentation files.
It is needed because assertions now includes assertions_basic. This is
not the cleanest solution, but it fixes the issue right now.

Those nodoc are required to hide the declarations and definitions that
are unintentionally introduced while documenting packages (which may not
necessarily introduce the 'assertions' package in the code that use
them).

Src-commit: c0ce24258d7ccfa6fae85fd1d776eb53387f9705