(core) Fix for compmod_skip bug (see below for details)
The predicate c_itf_internal:do_load_compilation_module/3 skipped
loading compilation modules that were already processed. This is not
correct when the module was processed but not (dynamically) loaded in
the context of c_itf_internal. This commit patches the predicate to
check that the interface of c_itf_internal has actually been extended to
include the dynamic load.

Src-commit: d5f35e61d433629773ae451ae5e16920384ef034