(core) Optional out-of-source builds on build/ directories
Out-of-source builds is a common practice in may build systems, which
place compiler intermediate outputs (.po, .itf, etc. in our case) in a
common separate build directory.

This commit replaces the experimental CIAOCACHEDIR=Dir enviroment
varible by a more convenient CIAOCCACHE variable. When CIAOCCACHE=1,
Ciao will automatically place the compiler outputs in the build/cache/
directory of the bundle's workspace.

For compatibility, compilation of modules that do not belong to any
bundle is performed in-source.

NOTE: CIAOCCACHE is still experimental in this commit, but it will be
the default after some initial testing period.

Src-commit: e1baa505c9bb0cce2c291e5b47605844d7fdcf6d