(core) avoid invalid cross-device link errors in file_buffer.pl
The temporary files where created in /tmp using mktemp_in_tmp/2.  This
patch creates the temporary file in the same directory where the final
output is expected so that we do not have errno=18 (invalid
cross-device link) problems in the C rename() function. This happens,
for example, when the /tmp is in a different device than the
target (e.g., 'tmpfs').

Src-commit: 39661c4ec0a5085a9f95d03b65b9e1254156413f