(core) Partial fix for spurious choicepoints in multifile predicates (see below)
See 'multifile_chpt' bug entry for more details about his bug. This
patch fixes an issue that causes that, under some circumstances the
following program:

  :- module(a, ...)
  :- multifile p/1.
  p(a).

  :- module(b, ...)
  :- multifile p/1.

needs a choicepoint for a p(X), despite there exists a unique clause
for p/1.

This patch delays the creation of wrappers for multifile predicates
until actual clauses are loaded. Note that this is not the optimal
solution (it will be backported from optim-comp).

Src-commit: 118a853ea3d06b3f8bcfd3fd048ac2632d1191cc