(core) Thread-safe and more efficient reimplemenation of exceptions
This new implementation uses (backtrackable) global variables to setup
the context for throwing and catching exceptions. The cost of
catch/3 is now constant (we do not copy the handler anymore), while
the cost of throw/1 is still linear in the size of the error term
(although this is somehow necessary to preserve it across backtracking
to reach the exception handler code).

Notes:

 - Since global variables are private per thread, this mechanism works
   fine in the presence of multiple engine workers.

 - We still use thrown/2 to send/receive errors terms but now it is a
   'concurrent' fact and it contains the thread ID as a second
   argument, to make it thread-safe.

 - To avoid a hard dependency to the concurrency.pl module, the code
   uses ':- import' and does a (cheap) runtime check on the
   availability of the code. It is safe to assume a thread ID equal to
   0 if concurrency is not (yet) loaded.

Src-commit: 90137e9d63c98a7eabf4ae6b696f85f9882a06c0