fixex bug in rtchecks for properties that deal with exceptions (no_exception/1, exception/2, not_fails...)
Summary:
The handler of some comp properties, including not_fails, sends an
rtcheck error when they catch an exception, that says something
like "not_fails property violated because exception X was thrown".
However, they also used to do that for the rtcheck exceptions themselves,
so rtcheck errors for inner calls were not reported as such, but as
exceptions that caused an rtcheck error in outer calls. See T350 for more
details

The fix makes the handler of those properties just ignore and rethrow the
caught exception if it is an rtcheck error

Test Plan: N/A

Reviewers: jfran

Reviewed By: jfran

Subscribers: jfran, ignaciocasso

Maniphest Tasks: T350

Differential Revision: http://phabricator.ciao-lang.org/D225

Src-commit: 96b0eec34387893ae18885c52d47d235ef223d4f