(core) New predicate abstractions, fix shared variables
Currently a predicate abstraction in a call like:

  foo((''(...) :- ...), ...)

may share some variables with the body, even if the "ShVs ->"
annotations is not used. These shared variables share in "one
direction" (during construction of the closure) and are copied in each
closure call.

This commit clarifies the semantics of predicate abstractions
such that only variables specified in ShVs:

  foo((ShVs -> ''(...) :- ...), ...)

will share with the caller's body variables. No other variables will
be implicitly shared.

Note that a call like:

  C = (ShVs -> ''(...) :- ...), foo(C, ...)

will force "meta-expansion", so runtime values of variables will be
treated during expansion. In this case there is no better way to avoid
sharing than using distinct variable names.

Src-commit: ca26e18d23e21c589bc07591ced961607193d8d0