(core) merge tokenize and read modules from OC
The tokenize and read modules in OC can read curly blocks as a list of sentences (including line numbers and variables), while the main version read curly blocks as lists of terms. This merge unifies both versions (selected with conditional code, although the OC version will be kept in the future).

Src-commit: 1f9ef8a47cdfa426c0036ad196ea9ac5ca81a0d6