(core) Removed negation of properties in rtchecks instrumentation
rtchecks instrumentation wrote the properties to be checked
negated, as in the following example:

```
:- success p(X,Y) => (int(X), list(Y)).

p(_,_).
```

which was translated to:

```
p(X,Y) :-
    ...
    'p/2$rtc1'(X,Y)
    ...
    check(...,[\+'basic_props_rtc:rtc_int'(X),'rtchecks_rt:non_inst'(list(Y),[Y])],...),
    ....
'p/2$rtc1'(X,Y).
```
While it makes sense on runtime to have those properties negated,
there was no need to complicate the instrumentation and the
rtchecks code by introducing that negation so early. Instead, in
this commit we make checkc/n handle that, and translate instead
to:

```
    ...
    check(...,['basic_props_rtc:rtc_int'(X),'rtchecks_rt:rtc_inst'(list(Y),[Y])],...),
    ....
```

Additionally, this commit fixes a bug in the runtime-check of
compat(nonvar, X) / compat(nonvar(X)), which should always
succeed but previously failed when X was a free variable.

Src-commit: bd40a5e86a45eaebe05eebcff284ee720499cbc8