(rtchecks) simplified instrumentation that check properties at runtime
Given a set of properties, the rtchecks instrumentation needs to
generate goals that check at runtime that those properties hold.  This
transformation includes using the alternative rtcheck version
(rtc_impl, if declared) or wrapping predicates in generic
instantiation checks (that makes sure that no variables are further
instantiated on the property success, etc.).

This translation was not working for program point assertions, and it
was duplicated in different parts of the instrumentation code, for
different kind of assertions.

This commit unifies and simplifies this code above in a single
get_check_props/4 predicate (which also fixes the bug in program-point
assertions).

Src-commit: 85c75f96ac4c8697058369cb423286411278bb40