(unittest) one-line output design for each test
First step towards redesigning unittest output. The output for each
test will be a single concise line describing the result, plus
optional details with different levels of verbosity. This commit deals
with that first line.

In this commit, that line has been changed to:
```
Note: (lns LB-LE) STATUS: p/a "Test Comment". Description.
```
where STATUS can be PASSED, FAILED, WARNING or ABORTED, and
Description is just an optional few words explanation of the warning
or reason it aborted.

Since a single test can have multiple results, due to the use of texec
properties times(N1), try_sols(N2), ... with Nj > 1, there was need to
decide a method to summarize all those results into one. For that, we
have assigned different priorities to the results, so that the final
result is the one with highest priority. E.g., if the test runner
ended up aborting, that will be the result written, regardless whether
it had time to successfully test some solutions and save their results
before aborting. The following list sorts the possible results from
higher to lower priority, together with their respective STATUS:

- Test runner aborted -> ABORTED
- Generation of test case from calls field failed -> WARNING
- Generation of test case from calls field aborted -> ABORTED
- Postcondition checking aborted -> ABORTED
- Timeout -> ABORTED
- rtcheck errors -> FAILED
- exceptions in goal, but no `not_fails` or `no_exception` comp
  properties -> WARNING
- goal failed, but no failure properties in the test assertion -> WARNING
- everything ok -> PASSED

Src-commit: 91a82e831be368b2fce44988e6ee5307aee7c61a