(engine) merged code for heap limit based events (wake, overflow, C-c)
WARNING: This makes CALLPAD static (rather than a dynamic value based
on Atom_Buffer_Length, which is an upper limit to the largest atom).

Dynamic CALLPAD does not seem to be necessary for any C builtin.
Moreover, since the atom table is global, heap limits should be
recomputed for every engine (not only for the current one).

Src-commit: 1c9d2b5a76cd336e87fe32dccbd17756b455d3e9