(core) Faster dependency checks in the compiler
This commit implements a faster mechanism for dependency checking in
c_itf.pl. Note that the current code is disabled (uncomment
":- compilation_fact(itf_sections)." to enable it) until we are
sure that it is safe.

It works by spliting itf data in two sections (level 0 and level 1).
Level 0 contains file dependencies and no actual interface data
(except for a few multifile declarations that are still required due
to a dynamic module load issue). Level 1 contains all the interface
data. Dependecy checking is done at level 0, which is very fast. When
needed, the compiler reads the rest of the itf data.

Details:

 - already_have_itf/2 extended with current ItfLevel

 - time_of_itf_data/3 extended with current ItfLevel

 - process_files_from/7 loads level 1 for the current file and level 0
   for dependencies

 - read_itf/6 reloads data if required ItfLevel increases

 - do_read_itf/5 accepts expected ItfLevel and an update time option
   (to distinguish itf level updates from new reads), needed for
   read_itf/6 and new_file_status/6

 - added get_file_deps/2, which reads up to level 0

 - added itf_handle_related_files/1 which completes
   handle_related_files/2 with level 1. This is called only if needed
   in do_treat_action/7 (when no recompilation is needed, savings are
   huge)

 - maintanance of time_of_itf_data/3 moved to read_record_file/4 and
   generate_itf/3

 - read_itf_data_of/3 stops early when required level is 0

 - itf_data ordered to save level 0 data (dependencies) before level 1
   (actual interface)

 - changed_dependencies/2 based purely on timestamps

Other:

 - Added a (fixed) bug entry for missing multifile data with sepearate
   deps itf section.

 - minor change in deepfind (use local always/1 instead of true/1)

Src-commit: 150a4c237ca65408dc03d46dba9888bd9944e333