(core) recover trans optimization in p_unit
The optimization (described in the module documentation) limits the
recursive loading of modules when they do not export any properties.
Imported modules where processed in two modes: direct (direct import)
or trans (transitive or indirect import). Indirect imports continued
loading imported modules only if those included more property
definitions.

This optimization was (accidentally?) disabled when the 'library
cache' code was introduced (due to a bug in the implementation). This
commits fixes it by generating .ast files that are independent on the
processing mode and moves outside the logic to compute the related
files.

The optimization disabled by default, uncomment the 'use_trans_opt'
flag to enable it.

Src-commit: 48d4bc4da259a39730b8ddc079c5b9a66bc2822d